Separate `routes.js` according to domains
